+++ uamqp/async_ops/mgmt_operation_async.py
@@ -78,7 +78,7 @@ class MgmtOperationAsync(MgmtOperation):
         :type message: ~uamqp.message.Message
         :param timeout: Provide an optional timeout in milliseconds within which a response
          to the management request must be received.
-        :type timeout: int
+        :type timeout: float
         :rtype: ~uamqp.message.Message
         """
         start_time = self._counter.get_current_ms()
